refactor: dissolve services layer into explicit architecture
