+++ CruiserJumpPractice/Presentation/StartupInitializer.cs
